Rename memdisk.rs to persist.rs
