+++ storage/src/node/mod.rs
@@ -34,8 +34,8 @@ use std::vec;
 
 pub mod branch;
 mod leaf;
-pub mod memdisk;
 pub mod path;
+pub mod persist;
 
 pub use branch::{BranchNode, Child};
 pub use leaf::LeafNode;
